requirements.txt
Flask>=2.2
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
gunicorn
